DAH-2389 fix duplicates
